+++ Changelog.md
@@ -13,6 +13,8 @@
 * Added support for php 7.1
 * Added Twig 2 support
 
+* DefaultPhpFileExtractor can be extended to define domain where should be extracted messages.
+
 ### 1.3.0 to 1.3.1
 
 * Fixed new messages not showing at the top in WebUI when in XLIFF format.